requirements.txt
fastapi==0.115.2
uvicorn==0.22.0
sqlalchemy==2.0.0
asyncpg==0.29.0 
aiohttp==3.10.10 
passlib==1.7.4
python-jose==3.3.0
aiogram==3.13.1
alembic==1.10.0
redis==4.5.0
celery==5.2.7
pydantic==2.9.2
python-dotenv
aiopg
jinja2
python-multipart==0.0.12
uvicorn[standard]

